requirements.txt
httpx==0.26.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.36
apscheduler==3.10.4
python-dateutil==2.8.2
nepali_datetime==1.0.8.4
python-dotenv==1.0.0
backoff==2.2.1
PyMySQL==1.1.1

